jobhunter/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cloudinary==1.44.1
Django==6.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
PyMySQL==1.1.2
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
